Discount added to Cashier

+++ Ubrania_Nowy/Ubrania_ASP.NET_Nowy/Models/Cloth.cs
@@ -15,7 +15,7 @@ namespace Ubrania_ASP.NET_Nowy.Models
         public string Type { get; set; } 
         public string Description { get; set; } 
         public int Price { get; set; }
-        public int Price_RL { get; set; }
+        public double Price_RL { get; set; }
         public int PriceCounter { get; set; }
         public int Agreement_Id { get; set; }
         public bool Sold { get; set; }
